pyproject.toml
[tool.poetry]
name = "customer-service"
version = "0.1.0"
description = "Customer Service Agent for Google Cloud Vertex AI"
authors = ["Audisoft"]
packages = [
    { include = "customer_service" }
]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
google-cloud-aiplatform = {extras = ["adk", "agent_engines"], version = "^1.88.0"}
google_api_python_client = "2.160.0"
protobuf = "6.30.2"
pydantic = "2.11.4"
pydantic-settings = "2.9.1"
python-dotenv = "1.1.0"
pytz = "2024.2"
cloudpickle = "3.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-mock = "^3.14.0"
pytest-cov = "^6.0.0"
scikit-learn = "^1.6.1"
pytest-asyncio = "^0.25.3"
flake8-pyproject = "^1.2.3"
pyink = "^24.10.1"
google-cloud-aiplatform = {extras = ["evaluation"], version = "^1.88.0"}

[tool.pytest.ini_options]
console_output_style = "progress"
addopts = "-vv -s"
testpaths = ["tests/", "eval/"]
markers = [
    "unit"
]
log_level="ERROR"
log_cli=false
log_auto_indent=true
log_cli_date_format="%Y-%m-%d %H:%M:%S"
log_cli_format ="[%(asctime)s] %(levelname)s (%(funcName)s) \t [%(pathname)s:%(lineno)d] %(message)s"
filterwarnings = [
    "ignore::UserWarning",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pyink]
line-length=80
pyink-indentation=4
pyink-use-majority-quotes = true